Fixed bug in details.php that’d use int val in updates values of commodities and cryptocurrencies

+++ details.php
@@ -56,7 +56,7 @@
 
             while($u = mysqli_fetch_array($res_updates)) {
                 $t = $u['time'];
-                $c = intval($u['current']);
+                $c = floatval($u['current']);
 
                 $updates_time[] = $t;
                 $updates_values[] = $c;
@@ -120,7 +120,7 @@
 
             while($u = mysqli_fetch_array($res_updates)) {
                 $t = $u['time'];
-                $c = intval($u['current']);
+                $c = floatval($u['current']);
 
                 $updates_time[] = $t;
                 $updates_values[] = $c;
